backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
google-generativeai==0.3.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pydantic==2.5.0
gunicorn==21.2.0 